move cleanup tags to a new module
