+++ crates/thetawave_interface/src/states.rs
@@ -1,4 +1,4 @@
-use bevy_ecs_macros::{Component, States};
+use bevy_ecs_macros::States;
 
 // states of the game
 #[derive(Debug, Clone, PartialEq, Eq, Hash, Default, States)]
@@ -20,24 +20,3 @@ pub enum GameStates {
     Playing,
     Paused,
 }
-
-#[derive(Component)]
-pub struct MainMenuCleanup;
-
-#[derive(Component)]
-pub struct GameCleanup;
-
-#[derive(Component)]
-pub struct GameOverCleanup;
-
-#[derive(Component)]
-pub struct VictoryCleanup;
-
-#[derive(Component)]
-pub struct PauseCleanup;
-
-#[derive(Component)]
-pub struct InstructionsCleanup;
-
-#[derive(Component)]
-pub struct CharacterSelectionCleanup;
