+++ src/background/mod.rs
@@ -9,8 +9,8 @@ use ron::de::from_bytes;
 use serde::Deserialize;
 use std::ops::Range;
 use thetawave_interface::{
+    cleanup_tags::GameCleanup,
     run::{RunDefeatType, RunEndEvent, RunOutcomeType},
-    states::GameCleanup,
 };
 use thiserror::Error;
