+++ src/spawnable/projectile/mod.rs
@@ -3,8 +3,8 @@ use bevy_rapier2d::prelude::*;
 use serde::Deserialize;
 use std::{collections::HashMap, string::ToString};
 use thetawave_interface::{
+    cleanup_tags::GameCleanup,
     spawnable::{ProjectileType, SpawnableType},
-    states::GameCleanup,
 };
 
 use crate::{
